Use roles correctly
Turns out the `menu` role is intended for a menu-item container, not the thing that opens the container.
The top-level items in a menu bar are counted as menu items, meaning the menu bar is a subset of `menu`, essentially.
Fortunately the changes are pretty easy, and all the functionality still works.